renaming TileViewPostList
