+++ src/components/PostList/ThumbNailPostList/ThumbNailPostList.tsx
@@ -1,6 +1,6 @@
 import { TilePostProps } from 'types';
 import TileViewItem from './TileViewItem';
-import styles from './TileViewPostList.module.scss';
+import styles from './ThumbNailPostList.module.scss';
 
 export type ViewType = 'small' | 'large';
 
@@ -10,17 +10,17 @@ interface Props {
   postList: TilePostProps[];
 }
 
-const TilePostList: React.FC<Props> = ({ listTitle, postList, viewType = 'small' }) => {
+const ThumbNailPostList: React.FC<Props> = ({ listTitle, postList, viewType = 'small' }) => {
   return (
     <div className={styles.TilePostList}>
       {listTitle && <h2 className={styles.listTitle}>{listTitle}</h2>}
       <div className={viewType === 'small' ? styles.smallWrapper : styles.largeWrapper}>
         {postList.map((value, index) => {
-          return <TileViewItem value={value} key={index} viewType={viewType} />;
+          return <TileViewItem value={value} key={index} />;
         })}
       </div>
     </div>
   );
 };
 
-export default TilePostList;
+export default ThumbNailPostList;
