+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import TilePostList from '@components/PostList/TileViewPostList/TileViewPostList';
+import ThumbNailPostList from '@components/PostList/ThumbNailPostList';
 import DefaultPostList from '@components/PostList/DefaultPostList/DefaultPostList';
 import HomeLayout from '@components/layout/HomeLayout';
 import styles from '@components/layout/HomeLayout/HomeLayout.module.scss';
@@ -8,7 +8,7 @@ const HomePage = () => {
   return (
     <HomeLayout>
       <div className={styles.tile}>
-        <TilePostList listTitle={'인기'} postList={HomePageTileMock} />
+        <ThumbNailPostList listTitle={'인기'} postList={HomePageTileMock} />
       </div>
       <div className={styles.default}>
         <DefaultPostList listTitle={'해외 주식'} postList={HomePageListMock} />
